requirements.txt
streamlit==1.31.1
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.8
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
python-dotenv==1.0.0
scipy==1.11.4
requests==2.31.0
openpyxl==3.1.2
reportlab==4.0.7
kaleido==0.2.1